+++ DerydocaEngine/TexturePacker.h
@@ -16,7 +16,7 @@ public:
 	TexturePacker();
 	~TexturePacker();
 
-	void addImage(int id, float sizeX, float sizeY, float bearingX, float bearingY, float advanceX, float advanceY, unsigned char* imageBuffer, int width, int height);
+	void addImage(unsigned long id, float sizeX, float sizeY, float bearingX, float bearingY, float advanceX, float advanceY, unsigned char* imageBuffer, int width, int height);
 	void packImages();
 	bool getIsDirty() const { return m_isDirty; }
 	Texture* allocTexture() { return dest.allocTexture(); };
@@ -28,7 +28,7 @@ public:
 private:
 	const int IMAGE_SIZE = 512;
 	vector<TexturePackerImage> m_images;
-	map<int, unsigned char*> m_imageBuffers;
+	map<unsigned long, unsigned char*> m_imageBuffers;
 	bool m_isDirty;
 	TexturePackerTextureData dest;
 	vector<IntRectangle> m_imageBounds;
